+++ py-polars/tests/unit/expr/test_exprs.py
@@ -430,34 +430,6 @@ def test_logical_boolean() -> None:
         df.select([(pl.col("a") > pl.col("b")) or (pl.col("b") > pl.col("b"))])
 
 
-# https://github.com/pola-rs/polars/issues/4951
-def test_ewm_with_multiple_chunks() -> None:
-    df0 = pl.DataFrame(
-        data=[
-            ("w", 6.0, 1.0),
-            ("x", 5.0, 2.0),
-            ("y", 4.0, 3.0),
-            ("z", 3.0, 4.0),
-        ],
-        schema=["a", "b", "c"],
-    ).with_columns(
-        [
-            pl.col(pl.Float64).log().diff().name.prefix("ld_"),
-        ]
-    )
-    assert df0.n_chunks() == 1
-
-    # NOTE: We aren't testing whether `select` creates two chunks;
-    # we just need two chunks to properly test `ewm_mean`
-    df1 = df0.select(["ld_b", "ld_c"])
-    assert df1.n_chunks() == 2
-
-    ewm_std = df1.with_columns(
-        pl.all().ewm_std(com=20).name.prefix("ewm_"),
-    )
-    assert ewm_std.null_count().sum_horizontal()[0] == 4
-
-
 def test_lit_dtypes() -> None:
     def lit_series(value: Any, dtype: pl.PolarsDataType | None) -> pl.Series:
         return pl.select(pl.lit(value, dtype=dtype)).to_series()
