+++ py-polars/tests/unit/series/test_series.py
@@ -1909,176 +1909,6 @@ def test_trigonometric_invalid_input() -> None:
         s.cosh()
 
 
-def test_ewm_mean() -> None:
-    s = pl.Series([2, 5, 3])
-
-    expected = pl.Series([2.0, 4.0, 3.4285714285714284])
-    assert_series_equal(s.ewm_mean(alpha=0.5, adjust=True, ignore_nulls=True), expected)
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=True, ignore_nulls=False), expected
-    )
-
-    expected = pl.Series([2.0, 3.8, 3.421053])
-    assert_series_equal(s.ewm_mean(com=2.0, adjust=True, ignore_nulls=True), expected)
-    assert_series_equal(s.ewm_mean(com=2.0, adjust=True, ignore_nulls=False), expected)
-
-    expected = pl.Series([2.0, 3.5, 3.25])
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=False, ignore_nulls=True), expected
-    )
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=False, ignore_nulls=False), expected
-    )
-
-    s = pl.Series([2, 3, 5, 7, 4])
-
-    expected = pl.Series([None, 2.666667, 4.0, 5.6, 4.774194])
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=True, min_periods=2, ignore_nulls=True), expected
-    )
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=True, min_periods=2, ignore_nulls=False), expected
-    )
-
-    expected = pl.Series([None, None, 4.0, 5.6, 4.774194])
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=True, min_periods=3, ignore_nulls=True), expected
-    )
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=True, min_periods=3, ignore_nulls=False), expected
-    )
-
-    s = pl.Series([None, 1.0, 5.0, 7.0, None, 2.0, 5.0, 4])
-
-    expected = pl.Series(
-        [
-            None,
-            1.0,
-            3.6666666666666665,
-            5.571428571428571,
-            5.571428571428571,
-            3.6666666666666665,
-            4.354838709677419,
-            4.174603174603175,
-        ],
-    )
-    assert_series_equal(s.ewm_mean(alpha=0.5, adjust=True, ignore_nulls=True), expected)
-    expected = pl.Series(
-        [
-            None,
-            1.0,
-            3.666666666666667,
-            5.571428571428571,
-            5.571428571428571,
-            3.08695652173913,
-            4.2,
-            4.092436974789916,
-        ]
-    )
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=True, ignore_nulls=False), expected
-    )
-
-    expected = pl.Series([None, 1.0, 3.0, 5.0, 5.0, 3.5, 4.25, 4.125])
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=False, ignore_nulls=True), expected
-    )
-
-    expected = pl.Series([None, 1.0, 3.0, 5.0, 5.0, 3.0, 4.0, 4.0])
-    assert_series_equal(
-        s.ewm_mean(alpha=0.5, adjust=False, ignore_nulls=False), expected
-    )
-
-
-def test_ewm_mean_leading_nulls() -> None:
-    for min_periods in [1, 2, 3]:
-        assert (
-            pl.Series([1, 2, 3, 4])
-            .ewm_mean(com=3, min_periods=min_periods)
-            .null_count()
-            == min_periods - 1
-        )
-    assert pl.Series([None, 1.0, 1.0, 1.0]).ewm_mean(
-        alpha=0.5, min_periods=1
-    ).to_list() == [None, 1.0, 1.0, 1.0]
-    assert pl.Series([None, 1.0, 1.0, 1.0]).ewm_mean(
-        alpha=0.5, min_periods=2
-    ).to_list() == [None, None, 1.0, 1.0]
-
-
-def test_ewm_mean_min_periods() -> None:
-    series = pl.Series([1.0, None, None, None])
-
-    ewm_mean = series.ewm_mean(alpha=0.5, min_periods=1)
-    assert ewm_mean.to_list() == [1.0, 1.0, 1.0, 1.0]
-    ewm_mean = series.ewm_mean(alpha=0.5, min_periods=2)
-    assert ewm_mean.to_list() == [None, None, None, None]
-
-    series = pl.Series([1.0, None, 2.0, None, 3.0])
-
-    ewm_mean = series.ewm_mean(alpha=0.5, min_periods=1)
-    assert_series_equal(
-        ewm_mean,
-        pl.Series(
-            [
-                1.0,
-                1.0,
-                1.6666666666666665,
-                1.6666666666666665,
-                2.4285714285714284,
-            ]
-        ),
-    )
-    ewm_mean = series.ewm_mean(alpha=0.5, min_periods=2)
-    assert_series_equal(
-        ewm_mean,
-        pl.Series(
-            [
-                None,
-                None,
-                1.6666666666666665,
-                1.6666666666666665,
-                2.4285714285714284,
-            ]
-        ),
-    )
-
-
-def test_ewm_std_var() -> None:
-    series = pl.Series("a", [2, 5, 3])
-
-    var = series.ewm_var(alpha=0.5)
-    std = series.ewm_std(alpha=0.5)
-
-    assert np.allclose(var, std**2, rtol=1e-16)
-
-
-def test_ewm_param_validation() -> None:
-    s = pl.Series("values", range(10))
-
-    with pytest.raises(ValueError, match="mutually exclusive"):
-        s.ewm_std(com=0.5, alpha=0.5)
-
-    with pytest.raises(ValueError, match="mutually exclusive"):
-        s.ewm_mean(span=1.5, half_life=0.75)
-
-    with pytest.raises(ValueError, match="mutually exclusive"):
-        s.ewm_var(alpha=0.5, span=1.5)
-
-    with pytest.raises(ValueError, match="require `com` >= 0"):
-        s.ewm_std(com=-0.5)
-
-    with pytest.raises(ValueError, match="require `span` >= 1"):
-        s.ewm_mean(span=0.5)
-
-    with pytest.raises(ValueError, match="require `half_life` > 0"):
-        s.ewm_var(half_life=0)
-
-    for alpha in (-0.5, -0.0000001, 0.0, 1.0000001, 1.5):
-        with pytest.raises(ValueError, match="require 0 < `alpha` <= 1"):
-            s.ewm_std(alpha=alpha)
-
-
 def test_product() -> None:
     a = pl.Series("a", [1, 2, 3])
     out = a.product()
